backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv
gunicorn
flask-mail